fixed minor issues

+++ contracts/Arth/Pools/ArthPool.sol
@@ -348,7 +348,7 @@ contract ArthPool is AccessControl {
         ARTH.permit(
             msg.sender,
             address(_spender),
-            uint256(-1)//amountToStake,
+            uint256(int256(-1)), //amountToStake,
             block.timestamp,
             v,
             r,
@@ -409,7 +409,7 @@ contract ArthPool is AccessControl {
         ARTH.permit(
             msg.sender,
             address(_spender),
-            uint256(-1),
+            uint256(int256(-1)),
             block.timestamp,
             v,
             r,
@@ -497,7 +497,7 @@ contract ArthPool is AccessControl {
         ARTH.permit(
             msg.sender,
             address(_spender),
-            uint256(-1),
+            uint256(int256(-1)),
             block.timestamp,
             v,
             r,
@@ -793,7 +793,7 @@ contract ArthPool is AccessControl {
         ARTHS.permit(
             msg.sender,
             address(_spender),
-            uint256(-1),
+            uint256(int256(-1)),
             block.timestamp,
             v,
             r,
